+++ src/Main/Props.ts
@@ -1,5 +1,0 @@
-import { MainState } from "./MainState";
-
-export interface AppProps {
-    mainState: MainState
-};