+++ src/Main/store/index.ts
@@ -1,10 +1,10 @@
 import { createStore, applyMiddleware, Store } from "redux";
 import thunkMiddleware from "redux-thunk"
-import { State } from "../../Shared/State";
+import { AppState } from "../../Shared/AppState";
 import { composeWithDevTools } from 'redux-devtools-extension';
 import { Reducer } from "../../Shared/reducers";
 
-const configureStore = (initialState?: State): Store<State | undefined> => {
+const configureStore = (initialState?: AppState): Store<AppState | undefined> => {
     const middlewares: any[] = [thunkMiddleware];
     const enhancer = composeWithDevTools(applyMiddleware(...middlewares))
     return createStore(Reducer, initialState, enhancer);