+++ src/Shared/State.ts
@@ -1,7 +1,0 @@
-import { MainState } from "../Main/MainState";
-import { FriendState } from "../Friends/FriendState";
-
-export interface State {
-    mainState: MainState,
-    friendState: FriendState
-};