+++ src/Shared/reducers.ts
@@ -1,9 +1,11 @@
 import { combineReducers } from "redux";
 import { friendReducer } from "../Friends/reducers/friendReducer";
 import { mainReducer } from "../Main/reducers/mainReducer";
-import { State } from "./State";
+import { AppState } from "./AppState";
+import { loginReducer } from "../Login/reducers/loginReducer";
 
-export const Reducer = combineReducers<State>({
+export const Reducer = combineReducers<AppState>({
     friendState: friendReducer,
-    mainState: mainReducer
+    mainState: mainReducer,
+    loginState: loginReducer
 });