+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 28/11/2019 - dunkybaldy
+* Props are immutable
+* State is mutable
+    * Use props to set state is class constructor
+* Configured BrowserRouter in base Root.tsx
+    * Must be as high level as possible
+    * any redirects done outside Root.tsx will not register correctly
+* Remove menu from default electron application
+* Example onClick login screen routes to Main.tsx
+
 ### 26/11/2019 - dunkybaldy
 * Props sent through to App class
 * Working application